updated the repo

+++ DevOps/clean_docker.sh
@@ -23,29 +23,37 @@ if [ $# -eq 0 ]; then
     exit 1
 
 elif [ $1 == '-l' ] || [ $1 == '--last' ]; then
-    echo
-    echo "Deleting last image created"
-    last=$(docker images | head -n 2 | tail -n 1 | awk '{print $3}')
-    docker rmi -f $last
-    # check if last image id deleted
-    if [ $? -eq 0 ]; then
-        echo "Last image deleted"
-    else
+    # ceck if there's any images
+    if [[ "$(docker images -q)" == "" ]]; then
         echo
-        echo "Last image not deleted"
+        echo "No images to delete"
+    else 
         echo
-        read -p "Do you want to stop and delete the container? [y/n] " ans
-        if [ $ans == 'y' ]; then
-            docker ps
-            echo
-            read -p "Enter container ID: " cont_id
-            docker stop $cont_id
-            docker rm $cont_id
-            docker rmi -f $last
+        echo "Deleting last image created"
+        last=$(docker images | head -n 2 | tail -n 1 | awk '{print $3}')
+        docker rmi -f $last
+        # check if last image id deleted
+        if [ $? -eq 0 ]; then
             echo
             echo "Last image deleted"
         else
+            echo
             echo "Last image not deleted"
+            echo
+            read -p "Do you want to stop and delete the container? [y/n] " ans
+            if [ $ans == 'y' ]; then
+                docker ps
+                echo
+                read -p "Enter container ID: " cont_id
+                docker stop $cont_id
+                docker rm $cont_id
+                docker rmi -f $last
+                echo
+                echo "Last image deleted"
+            else
+                echo
+                echo "Last image not deleted"
+            fi
         fi
     fi
